refactor(p2p): move GasLimitsValidator into its own file
Code move only, no behavior change. GasLimitsValidator no longer shares
ownership with GasTxValidator, so it gets its own gas_limits_validator.ts
and test file, matching the one-validator-per-file layout of the directory.
Exports still flow through the index barrel, so external imports are
unaffected.
